Change in encoding workflow of the file
The txt_to_txtd CLI now takes one argument, the name of the file to be encoded, in the input.
A folder is created with the same name as the file, and the encoded file is placed inside that folder.
A checksum file is also created in the same folder, containing the SHA256 checksum of the original file.
The encoded file is named as input_file_name.encoded.txtd.
The checksum file is named as input_file_name.checksum.txt.