Change in decoding workflow of the file
The txtd_to_txt CLI now takes one argument, the name of the folder in which the file to be decoded resides, in the input.
It then searches for the file named input_folder_name.encoded.txtd in the folder and decodes it to input_folder_name_decoded.txt.
It then generates the checksum of the decoded file and compares it with the one present in input_folder_name.checksum.txt.
If the checksums match, it prints a success message; otherwise, it prints an error message.

NOTE: the name of the folder and the name of the files before the .encoded.txtd and .checksum.txt extensions must be the same.